requirements.txt
# requirements.txt
groq
trafilatura
requests
lxml[html_clean]
lxml_html_clean
google-generativeai
python-dotenv
rich
langchain
sentence-transformers
chromadb
googlesearch-python
numpy
transformers
torch
beautifulsoup4
scikit-learn
newsapi-python
newspaper3k
lxml
cross-encoder

# source checker
python-whois
tldextract
dnspython
validators
feedparser

# language detection
langdetect


textstat
transformers
sentence-transformers
spacy
torch